requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
PyGithub==2.1.1
psycopg2-binary==2.9.9
playwright==1.40.0
pandas==2.1.4
sqlalchemy==2.0.23
anthropic==0.7.8
python-dateutil==2.8.2
Pillow==10.1.0

